Moving options.py OptionsDictionary class definition to util in order to avoid a circular import problem when it lived in core.
