+++ openmdao/core/__init__.py
@@ -4,6 +4,5 @@ from openmdao.core.parallel_group import ParallelGroup
 from openmdao.core.problem import Problem
 from openmdao.core.system import System
 from openmdao.core.driver import Driver
-from openmdao.core.options import OptionsDictionary
 from openmdao.core.basic_impl import BasicImpl
 from openmdao.core.vec_wrapper import VecWrapper, SrcVecWrapper, TgtVecWrapper
