+++ openmdao/util/__init__.py
@@ -1,3 +1,4 @@
 
 from openmdao.util.array_util import evenly_distrib_idxs, to_slices
 from openmdao.util.file_util import find_files, find_up
+from openmdao.util.options import OptionsDictionary
