+++ openmdao/util/options.py
